+++ tests/test_parser.py
@@ -1,7 +1,5 @@
 """Parser class unit tests."""
 
-from pathlib import Path
-
 import pytest
 
 from faddr.exceptions import FaddrParserUnknownProfile